Revert %%autosetup patch due to another regression
The repo may be initialized before the macro runs, in which case we
would fail to create the initial main branch and thus fail on "git
branch --set-upstream-to=main".  This caused an FTBFS in the edk2
package:

https://github.com/rpm-software-management/rpm/pull/2121#issuecomment-1202349789

Revert the patch for now as a hotfit and revisit the actual use case
later upstream, either fixing it properly or reverting there as well.